+++ app/src/main/java/com/ysshin/fine_dust_app/adapters/AppBindingAdapters.kt
@@ -1,16 +1,19 @@
 package com.ysshin.fine_dust_app.adapters
 
 import android.util.Log
-import android.view.View
+import androidx.constraintlayout.widget.ConstraintLayout
+import androidx.core.content.ContextCompat
 import androidx.databinding.BindingAdapter
-import com.ysshin.fine_dust_app.utils.DustState
 import com.ysshin.fine_dust_app.utils.ImageUtils
 
 object AppBindingAdapters {
     @JvmStatic
     @BindingAdapter("bind_background_image")
-    fun bindBackgroundImage(view: View, dustState: DustState?) {
-        Log.d("dust state", "${ImageUtils.convertStateToImage(dustState ?: DustState.GOOD)}")
-        view.setBackgroundResource(ImageUtils.convertStateToImage(dustState ?: DustState.GOOD))
+    fun bindBackgroundImage(layout: ConstraintLayout, dustState: Int) {
+        Log.d("dust state", "${ImageUtils.convertStateToImage(dustState)}")
+        layout.background = ContextCompat.getDrawable(
+            layout.context,
+            ImageUtils.convertStateToImage(dustState)
+        )
     }
 }
