--- app/src/main/java/com/ysshin/fine_dust_app/utils/DustState.kt
@@ -1,5 +1,0 @@
-package com.ysshin.fine_dust_app.utils
-
-enum class DustState {
-    GOOD, MODERATE, UNHEALTHY, VERY_UNHEALTHY
-}
